Improvements in the README.md
After noticing in some issues  that users were not grasping
the use of `./bootstrap` script prior to `./configure` when
building from the main branch  I added more information and
reestructured  headers/sections  in  an  attempt to make it
more evident.

I also renamed the "Installation"  section to "Building" so
it becomes more descriptive and highlights the fact that an
"Installation" section with instructions for different OSes
is needed (at least for the main systems/distributions).